Had to disable preprocessor warnings to allow build without filling my screen due to redefinitions in resource.h: See extended.
The resource.h VS2005 generated redefines a lot of things defined in
windows.h (and, I checked, defines them to be the same as what windows.h
defines them as). Until I'm done working on the dialogs it's going to
need to stay that way. After I am done, I can remove the definitions
that are being redefined. So as much as I'm uncomfortable with turning
off all preprocessor warnings, it'll have to be that way in the
experimental branch until I'm done with this bit.